Simplify the TypeScript plugin
