+++ packages/vscode-mdx/package.json
@@ -41,12 +41,16 @@
     "vscode:prepublish": "npm run build"
   },
   "devDependencies": {
+    "@mdx-js/language-service": "0.5.3",
     "@types/vscode": "^1.82.0",
     "@volar/language-server": "~2.1.0",
     "@volar/vscode": "~2.1.0",
     "@vscode/vsce": "^2.0.0",
     "esbuild": "^0.20.0",
-    "ovsx": "^0.8.0"
+    "load-plugin": "^5.0.0",
+    "ovsx": "^0.8.0",
+    "remark-frontmatter": "^5.0.0",
+    "remark-gfm": "^4.0.0"
   },
   "galleryBanner": {
     "color": "#fcb32c",
